Move tests to match folder structure
